root: move initial implementation to "pre_alpha" module
